cop/requirements.txt
# Auto COP Requirements
# Python 3.11+ required

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Data Validation
pydantic==2.8.2

# Image Processing (optional for now)
# rasterio==1.3.9  # Commented out as specified
# shapely==2.0.2   # Commented out as specified
# geopandas==0.14.1  # Commented out as specified

# JSON Schema Validation
jsonschema==4.23.0

# YAML Processing
PyYAML==6.0.2

# HTTP Client for testing
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# CORS Support (included in FastAPI)
python-multipart==0.0.12
